Initial skeleton code for C++ Implementation (#16)
* Add libcask native dll project to solution
* Add stub implementation aligned to test ICask interface
* Add P/Invoke interop in tests
* Use MSBuild and Xunit magic so that we
  1. Have a good experience in VS on Windows for C++
  2. Don't regress the C# dev experience x-plat for C#

The focus here is on getting the build off the ground so folks can work
in parallel. The API shape will surely change when an actual non-stub
C++ implementation is worked on.

This currently works only on Windows x64 and only with big VS, but
everything is set up there to the point where you can drop a breakpoint
in .cpp and hit it from the Test Explorer.

More work will be needed to match this experience for Linux, Mac, VS
Code on Windows, and non-x64 architectures.

There is also no PR validation coverage yet because that will require
infrastructure to build with big VS and/or solving x-plat.

### Also (not strictly related)
* Revert back to CRLF on Windows

There are too many issues with VS putting snippets with CRLF into files
and making them have mixed line endings.