Add standard BNF for 256-bit primary key. (#26)
This change adds a spare markdown standard BNF specification for CASK
keys.

The markdown additionally describes how to dereference/decompose an
encoded form of a key or its bytewise (decoded) form to obtain
constituent elements.

This change also adds a naive test helper that breaks down a 3-byte
sequence into 4 easily obtained 6-bit values (rendered as bytes). It is
not clear this code should be productized. This code is tested by a
brute force unit tests that takes over 30 seconds on .NET framework to
run (will be updated tomorrow to resolve this issue).