Add C3ID implementation (#21)
This implements C3ID computation that we will need soon. 

The implementation differs from the one in
https://github.com/microsoft/security-utilities and this seeks to become
the new standard.

Changes:
- There is no use of "Microsoft"
- There is no byte array to hex conversion in the intermediate
computation
- Optimized to avoid allocations
- Raw value is truncated to 12 bytes instead of 15
- "C3ID" is prefixed to the canonical representation in base64

Tests include a naive, unoptimized implementation that is easier to
understand, and we check that both implementations match. We also check
that we produce the same deterministic value for a sampling of
hard-coded test input.

This change also fixes edge cases with empty span input in encoding
polyfills.