C3ID addition to primary key + timestamp expansions. (#31)
This change first incorporates a 16-byte correlating id into a primary
key. By doing this, we ensure there is a non-sensitive component to
every key that can be used to collect/store telemetry (without requiring
that we process any sensitive data). Including this data also enables
another interesting security value: we can require that secure
user-experiences render only the non-sensitive portion of a key to
users, never putting a fully exploitable key into view. The C3ID itself,
as a 16-byte value, has sufficient uniqueness to never collide with any
other data value (that is generated in a proper way). 16 bytes comprises
128 bits of random data: a GUID itself only comprises 122 bits once you
account for its 6 bits of reserved data.

This change also attempts to integrate three other loose unresolved
feature requests:

- Providing sufficient timestamp data to mint/track short-lived secrets
- A mechanism to specify an expiry/lifetime for the secret
- A provider-specific key kind data value (which is the analog to the
CASK key kind)

The BNF has been updated to allow for everything above. As usual, the
data is organized in an attempt to align cleanly, as far as we can,
across 3-byte boundaries.

*NOTE*: the C3ID itself adds significantly to the length of the keys and
there is a fair amount of distributed padding in the key format to help
with alignment. Based on this, I've reverted to a previous decision to
only require 3 bytes of the CRC32 checksum - which was a point of
contention previously. At this point, we clearly do *not* require the
full 4-byte checksum for scan accuracy.

The chances of a false positive in the general format without accounting
for the checksum is 1 in 717,419,295,479. Adding the 24-bit checksum
brings odds to 1 in 12 quintillion. All of this is to say that the
checksum itself may be entirely dispensable. The value we would lose is
the ability to detect keys that users have either tweaked or simply
hand-authored. i.e., there is some friction in computation that is
arguably useful to avoid these patterns. But perhaps not. By dropping
the checksum requirement, for example, it becomes extremely simple to
author test keys (either those that use the reserved `TEST` provider
signature or those that simply recapitulate a single base64-encoded
character where we typically expect randomized data to exist).

One other note on this, all the math here only depends on the fixed
values for the general CASK case. If you account for a specific provider
signature (24 bits), accuracy again shoots in the quadrillions. We have
also defined a per-provider key kind that improves accuracy further.

Computing the chance of a false positive in randomized base64 encoded
bytes:
16/64 * (limited encodable chars due to reserved padding after random
data)
 1/64 * 'J'
 1/64 * 'Q'
 1/64 * 'Q'
 1/64 * 'J'
 3/64 * ('A', 'H' or 'I' in CASK key kind char)
4/64 * (reserved padding that allows for consistent c3id encoding)
12/64 * (restricted chars valid for timestamp month)
31/64 * (chars valid for timestamp day)
24/64 * (chars valid for timestamp hour)
60/64 (chars valid for timestamp minute)