Design revision direct encoding of correlating id zap hmac (#32)
We have recently completed a design review with some important outcomes:

1. Security review has deemed any post-exposure transformation of a
secret as breaking the 'single key, single use' principle. We have
accordingly dropped the `CaskComputedCorrelatingId` as a concept.
Instead, we propose to include a non-sensitive, unique identifier in
every minted CASK construct. We refer to this value as a `CASK
correlating id` or `c2id`, or simply a `correlating id`.
1. Because a correlating id for any minted secret is now directly
encoded in the key, it is available to extract and publish in other
contexts, for example, it may be used as a standalone parameter in a SAS
URL (providing a pointer to the underlying secret used to sign the SAS
HMAC) or placed contiguously with an HMAC. Because this data is easily
extracted from a key, we do not need the complexity of
proposing/implementing API for the HMAC scenario.
1. This removal also emphasizes another important emerging principle,
which is that it is useful to limit coding/design in our format as far
as possible to avoid cryptographic operations in the core format.
1. A parallel track exploring advanced detection scenarios has
demonstrated the utility of anchoring search around the fixed CASK
signature with some contiguous data that describes the size of all data
to the left of the signature.
1. Based on the increased size of a secret, to accommodate the
non-sensitive correlating id, and after examining the math around chance
of false positives occurring in other base64-encoded data, we have
decided to eliminate the CRC32 check.

Based on the above, literal changes in the code include the following:
- Update the primary key BNF and supporting code (including CASK regex)
to reflect a general format update that adds a key size and provider key
kind value, moves the CASK key kind to a different location, moves the
optional data to the very end of the format, adds a non-sensitive
(currently 16 byte) correlating id and removes a reserved byte.
- Not related to the API/cryptography simplification, the new format
adds a `minutes` character to the encoded timestamp as well as an 18-bit
expiry which represents a count of 5-minute increments, allowing for a
maximal key lifetime of ~900 days.
- Eliminate all API, tests and pseudo-code related to HMAC computation
and validation in both C# and C++.
- Eliminate code/test/const literals related to CRC32.
- Since we currently do retain a key kind representing HMAC, I have
renamed use of `hash` to `Hmac` to clarify the use case.
- I have renamed `KeyKind` to `CaskKeyKind` due to introducing a
provider-specific key kind.
- Because we now allow (and even expect) providers to specify a key
kind, I have introduced a `providerKeyKind` argument for this data with
appropriate validation/etc.
- [SORRY!] I have incidentally fixed a style issue that I happen to care
about, introducing a C++ style identation schema for parameters where
the first arg immediate follows the opening paren and all others indent
at that same column, on a new line.

Several open questions and other TODOs:
- The key size proposal in particular requires a leading padding byte.
This happens naturally for keys that are a multiple of 16, 32, or 64
bytes. 48 byte keys fall cleanly along a 3-byte boundary, however and
would require 3 bytes of padding to render the size argument. We may be
able to discuss space for the size key elsewhere that avoids this issue.
- Critically, our implementation doesn't handle any CASK productions of
varying size. If we're going to permit this, we should update the BNF,
API, examples, etc. One interesting question here is whether we allow
for 128-bit keys. These are considered unbreakable in that scenario
today, but 128 hashing algorithms are not. We need to decide whether 256
is a threshold for the format in general *or* accept complexity of
allowing 128-bits in some scenarios but not others. To maintain our
`stay out of cryptographic concerns` as far as possible, it may be
simplest to require 256 bits as a minimum.
- Open question: how far should we formalize a discrete/separable notion
of all the elements of a secret *other* than the randomized
bytes/hash/derived key that comprise the literally sensitive data? One
problem with separating them is that by eliminating the sensitive
component, a pattern is no longer exploitable and therefore not
necessarily exposed. And so there is at least one argument for advising
against this.
- There may be other adjustments to the literal format, specifically, if
we drop the c2id data to 15 bytes, we recover sufficient space in the
key to include the size element. Now, however, placement becomes an
issue (as having this character be contiguous to the core signature
pattern is helpful for advanced SIMD string search capabilities.
- To distinguish CASK from previous experimentation in this area, we
should use a new CASK specific fixed signature, of `QJJQ`.
- Because we did not close on the format in 2024, we should reset our
first specified year value to 2025 (and try not to let it slip to 2026).
- Our C++ implementation is still stubbed out.