Provider key kind as char. (#41)
Code change to resolve #33.

Currently, provider key kind is a string, which we check to ensure is
non-null, has length of 1, is url safe base64, etc.

The change is to accept a char value only for the provider key kind
(which ends up being encoded into 6 bits of the key). This is a more
appropriate data type and also simplifies validation and some other
logic.

---------

Co-authored-by: Nick Guerrera <nguerrera@microsoft.com>