Significant revision of cask. Remove expiry. Reorder data. Eliminate CASK kind. (#43)
This significant revision of CASK contains the following changes:

- Sensitive data key moved from its current location. As a result of
this move, we restore reserved padding bits for sensitive component
sizes of 16, 32 and 64 bytes. A 48-byte sensitive component aligns
cleanly along a 3-byte boundary and has no padding.
- The timestamp is moved to follow the CASK signature. Minutes data is
retained.
- The expiry is dropped from the core format. This data would be helpful
to responders to encode, as by reading the key it would be possible to
understand whether the key, as originally allocated was valid in some
exposure timeframe. A key CASK principle is for the core format to
publish data that is useful for static consumption and not tempting to
evaluate at runtime. And so we drop the expiry as too tempting for
runtime evaluation.
- The CASK key kind is eliminated. Various classes of key, primary key,
derived, HMAC can be specified in the provider-specific key kind
component.
- We now encode the optional data size, as a count of 3-byte optional
segments. By encoding this length as well as the sensitive component
length in characters at a fixed location relative to the CASK signature
`QJJQ` we enable high-performance detections that start with locating
the CASK signature in test/data streams.
- The correlating id is dropped to 15 bytes in length. In general, a
16-byte value (128 bits) is preferred in scenarios for a strong
guarantee of uniqueness. In practice, 120 bits is sufficient.

The BNF is renamed to `CaskSecrets.md` to reflect that there's no longer
a specific CASK key kind in play. The churn in the BNF was sufficient
that rather than renaming first so that we could diff, I simply created
a clean copy of the new BNF. But in case anyone finds it helpful to see
the explicit diffs from previous, I've left an edited version of the old
BNF behind to examine. This will be deleted, obviously, once everything
is reviewed/refined/accepted.

This change resolved #35, #36 and #37.

---------

Co-authored-by: Nick Guerrera <nguerrera@microsoft.com>