[WIP] Update some occurrences of sensitiveDataSize to secretSize to reflect enum name change. (#51)
I foolishly renamed the `SensitiveDataSize` enum to `SecretSize` in the
last PR, without appreciating the scale of rename required to put the
code in a consistent condition. We need terminology here that makes the
sensitive component data clear.

`SecretSize` == a count of 16-byte segments that are used to store the
literal secret encoded in a CASK secret (and now I'm getting concerned
that calling all of these cask 'secrets' is itself confusing).
`secretSizeInBytes` = the computed size in bytes for the literal secret.
`paddingSecretSizeInBytes` = the final, padded (if necessary) byte range
that brings the sensitive data component to a 3-byte boundary.