Convert to destination span pattern. (#54)
This change:

- Updates the key generating logic to follow a clean pattern of using
(and subsequently advancing) a destination Span<byte> to populate the
key.
- Defines many new limits/descriptive constants to replace embedded
magic numbers in source.
- Updates various source code comments to bring formatting into
conformance with the extension we use for this (rewrap).
- Several tests professing to test negative code paths in IsCask
actually verified exception conditions in `Cask.GenerateKey`. These
tests have been renamed to reflect the target method under test.
- Added several tests that modify keys to invalidate provider kind,
secret size and other data to verify that `Cask.IsCask` fails properly,
without exceptions or asserts.
- Added a specific test for an unusual corner case in the standard, viz.
that a key length of 90 bytes is not valid (as this length is only
possible for a 384-bit key but exceeds the permissible 12-byte limit on
optional data. This gap is caused by the fact that a 384-bit (48 byte)
key does not require padding to be 3-byte aligned but the next size,
64-bytes, does. An invalid key in this unusual condition satisfies the
general CASK regex, unfortunately, requiring an update to a test helper
to suppress verification of this for the new test. It may be possible we
can address this issue in the regex as part of the 11th hour tinkering
proposed in #46.
- Converted some tests to `[Theory]`, configuring them to test all
possible keys sizes (as opposed to the default size only, 256 bits).

This change resolves #47 and #49 and at least partially addresses #50.