Reorder sensitive data, optional data size, et al. as described in #48 (#55)
This PR establishes a new ordering of the raw 6-byte sequence that
incorporates the timestamp, sensitive data size, optional data size, and
provider key kind. Previously the timestamp preceded all other data, so
the practical description of the required change is to move this data to
precede the provider kind (leaving the sensitive data and optional data
sizes in the same order and now immediately following the CASK
signature).

#46 includes the detailed rationale for the change). Each encoded
character (of which there are 8) in this base64-encoded sequence is
interpreted independently, as follows and now in the following order:

1. Secret data size
2. Optional data size
3. Timestamp (viz., time-of-key-allocation) year
4. Timestamp  month
5. Timestamp  day
6. Timestamp  hour
7. Timestamp  minute
8. Provider key kind.

In addition to updating the BNF, pseudocode and API logic, this change
adds several missing tests to validate that illegal timestamp values are
not accepted as valid.

This change resolves #46 and #45. I have also opportunistically
continued to implement our destination `Span` pattern for accessing data
(in places not already covered by #49).