YAFC: Remove correlating id. Expand and move timestamp. (#62)
- This change removes the correlating id. On an internal security review
of this data, we do not believe it is useful to include as a standard
part of the format.
- We have expanded the timestamp to include seconds data and moved it to
the rightmost portion of the key. This both increases the granularity of
the timestamp (useful for short-lived secret scenarios) and also
effectively creates a non-sensitive identifier that can be used in error
reporting, etc., to describe a finding. In VS, for example, we would
truncate the secret and display only the rightmost six characters of the
key (i.e., the timestamp) in the error-list. The benefit of retaining a
small (non-sensitive) portion of the secret in a user-facing string is
that this data can be used to search the original scan target for the
exposure, can be used in email/chat communications about an exposure,
etc.
- We have sprinkled additional padding in the format that brings
detection accuracy to extremely high levels. This allows us to avoid
techniques such as incorporating a checksum into key
production/validation. A checksum is not inherently undesirable but we
are trying to minimize supporting API required to generate and validate
CASK secrets.
- I have added exhaustive tests that ensure all padding is required to
be present for a key to be deemed valid.

This PR resolves #60. We expressly have chosen *not* to move forward
with a fixed header/footer solution (as discussed in #48) at this time.
See that issue for deeper discussion on open questions.

---------

Co-authored-by: Nick Guerrera <nguerrera@microsoft.com>