Make IsCask and IsCaskUtf8 implementations consistent (#66)
`IsCask` taking UTF-16 validated the secret size up-front while
`IsCaskUtf8` deferred to `IsCaskBytes` to do so. I noticed this
discrepancy while looking at #64 where I am experimenting with tricks
for sharing code between UTF-8 and UTF-16 cases.

The fix here is to let `IsCaskBytes` handle it in both cases, which is
consistent with several other fields that are only validated after
decoding to bytes.

This change also moves `IsCaskUtf8` next to `IsCask` to emphasize that
these are the same operation for different encodings, to make it easier
to look for unwanted differences between them, and to keep public API
above helpers.

Finally, parameter names and method documentation are updated to help
communication the distinction between the binary decoded form and the
base64url-encoded textual forms.